TDKB-2132 TDK-B Build Failure Due to Warnings to Error Flag
Reason for change: Fixing the Typecasting issues in TDK-B Files
Test Procedure:TDKB Builds should be successful
Risks: None

(cherry picked from commit a372e084b77585049e876db3534d60c2e06c6298)

+++ tdkb_lib/src/ssp_main.c
@@ -68,7 +68,7 @@ int  cmd_dispatch(int  command)
     char*                           pParamNames[]      = {"Device."};
     parameterValStruct_t**          ppReturnVal        = NULL;
     int                             ulReturnValCount   = 0;
-    ULONG                           i                  = 0;
+    int                             i                  = 0;
     switch ( command )
     {
         case	'e' :
